+++ example/bin/prompt.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:flex/flex.dart';
 
 Future<void> main(List<String> arguments) async {
+  print('\n-------- Prompt Test --------');
   final name = await prompt<String>('What is your name?');
   print(r'Hello, $name'.fillTemplate({'name': name}));
   exit(0);